# docs/swagger.yaml
basePath: /
definitions:
  dto.TodoForm:
    properties:
      description:
        type: string
      status:
        enum:
        - PENDING
        - COMPLETED
        - PROGRESS
        type: string
      title:
        type: string
    required:
    - status
    - title
    type: object
  dto.UpdateStatusForm:
    properties:
      status:
        enum:
        - PENDING
        - COMPLETED
        - PROGRESS
        type: string
    required:
    - status
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is a sample server for managing todos.
  title: Todo API
  version: "1.0"
paths:
  /api/v1/todos:
    get:
      description: Get a list of all Todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Response'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: List all Todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new Todo with the given details
      parameters:
      - description: Todo form
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/dto.TodoForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create a new Todo
      tags:
      - todos
  /api/v1/todos/{id}:
    delete:
      description: Soft delete a Todo by setting the deleted_at field
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Soft delete a Todo by ID
      tags:
      - todos
    get:
      description: Get details of a Todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get a Todo by ID
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Update the details of a Todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo form
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/dto.TodoForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update an existing Todo
      tags:
      - todos
  /api/v1/todos/{id}/status:
    put:
      consumes:
      - application/json
      description: Update only the status field of a Todo by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: New status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStatusForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update the status of a Todo
      tags:
      - todos
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"